WIN32: add special windows compat import from utils
